+++ Thread_Run/src/kr/hs/dgsw/network/lesson2/Thread_interrupt.java
@@ -1,13 +1,13 @@
-package kr.hs.dgsw.network;
+package kr.hs.dgsw.network.lesson2;
 
-public class Thread1 extends Thread{
+public class Thread_interrupt extends Thread{
     public void run() {
 
         try {
             for(int i=0; i<500; i++) {
                 System.out.println("쓰레드 1 : "+i);
             }
-            Thread1.sleep(1);
+            Thread_interrupt.sleep(1);
         } catch (InterruptedException e){
             System.out.println("강제종료됨");
         }
