+++ Thread_Run/src/kr/hs/dgsw/network/lesson2/WhitSynchronizedMain.java
@@ -1,4 +1,4 @@
-package kr.hs.dgsw.network;
+package kr.hs.dgsw.network.lesson2;
 
 public class WhitSynchronizedMain {
