+++ Thread_Run/src/kr/hs/dgsw/network/lesson3/CakeEater.java
@@ -1,4 +1,4 @@
-package kr.hs.dgsw.network;
+package kr.hs.dgsw.network.lesson3;
 
 public class CakeEater extends Thread {
 	private CakePlate cake;
