+++ Thread_Run/src/kr/hs/dgsw/network/lesson3/CakeEatings.java
@@ -1,4 +1,4 @@
-package kr.hs.dgsw.network;
+package kr.hs.dgsw.network.lesson3;
 
 public class CakeEatings {
