+++ Thread_Run/src/kr/hs/dgsw/network/lesson3/CakePlate.java
@@ -1,4 +1,4 @@
-package kr.hs.dgsw.network;
+package kr.hs.dgsw.network.lesson3;
 
 public class CakePlate {
 	private int breadCount = 0;
