+++ Thread_Run/src/kr/hs/dgsw/network/lesson3/DiningPhilosopherMain.java
@@ -1,0 +1,4 @@
+package kr.hs.dgsw.network.lesson3;
+
+public class DiningPhilosopherMain {
+}
