+++ Thread_Run/src/kr/hs/dgsw/network/lesson4/Internat.java
@@ -1,0 +1,6 @@
+package kr.hs.dgsw.network.lesson4;
+
+public class Internat {
+    private String url = "https://open.neis.go.kr/hub/mealServiceDietInfo?Type=json&ATPT_OFCDC_SC_CODE=D10&SD_SCHUL_CODE=7240454&MLSV_YMD=20220503";
+    
+}
